requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
beautifulsoup4==4.14.3
bs4==0.0.2
camelot-py==1.0.9
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
cssselect==1.3.0
curl_cffi==0.14.0
distro==1.9.0
et_xmlfile==2.0.0
fake-useragent==2.2.0
fastapi==0.125.0
frozendict==2.4.7
h11==0.16.0
html5lib==1.1
idna==3.11
importlib_metadata==8.7.0
jpype1==1.6.0
lxml==6.0.2
lxml_html_clean==0.4.3
multitasking==0.0.12
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parse==1.20.2
pdfminer.six==20251107
pdfplumber==0.11.8
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.1
protobuf==6.33.2
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==11.1.1
pypdf==5.9.0
pypdfium2==5.2.0
pyppeteer==2.0.0
pyquery==2.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-html==0.10.0
six==1.17.0
soupsieve==2.8
starlette==0.50.0
tabula-py==2.10.0
tabulate==0.9.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.20
uvicorn==0.38.0
w3lib==2.3.1
webencodings==0.5.1
websockets==15.0.1
yfinance==0.2.66
zipp==3.23.0
